fs: move utility functions to internal/fs
PR-URL: https://github.com/nodejs/node/pull/18777
